requirements.txt
PyYAML
coverage
greeking
flake8
python-coveralls
six
tox
twine
django-appconf
